🚚 consolidate into events.ts
